+++ src/main/kotlin/com/github/sleepypanda/feesh/settings/categories/Items.kt
@@ -1,6 +1,7 @@
 package com.github.sleepypanda.feesh.settings.categories
 
 import com.github.sleepypanda.feesh.features.items.background.TrashBooksHighlighter
+import com.github.sleepypanda.feesh.features.items.tooltip.TooltipManager
 import com.github.sleepypanda.feesh.utils.enums.ColorCodes.*
 import com.github.sleepypanda.feesh.utils.enums.FormattingCodes.*
 import com.teamresourceful.resourcefulconfigkt.api.CategoryKt
@@ -58,5 +59,20 @@ object Items : CategoryKt("Items") {
     var showAutoRecombFlag by boolean(false) {
         this.name = Translated("Auto-recomb flag")
         this.description = Translated("Renders recomb upgrade flag (R) for auto-recombobulated fishing drops in the item slot.")
+    }
+
+    init {
+        separator {
+            this.title = "${AQUA}${BOLD}Tooltip"
+        }
+    }
+
+    var showExpertiseKillsTooltip by ObservableEntry(boolean(false) {
+        this.name = Translated("Expertise kills counter")
+        this.description = Translated("Shows exact expertise kills counter on the fishing rods.")
+    }) { prev, new ->
+        if (prev != new) {
+            TooltipManager.refreshEnabledAdders()
+        }
     }
 }